Add entity sync utility
