--- .wrangler/state/v3/r2/hassio-proxy-logs/blobs/47cba83938db27f3a7080fa336571d7275f3d00304619463af7e95cf030557cd00000198e19d9dec
@@ -1,1 +1,0 @@
-{"level":"error","message":"Test error log from diagnostics dashboard","timestamp":"2025-08-25T14:24:19.876Z"}
