--- .wrangler/state/v3/r2/hassio-proxy-logs/blobs/77b56101bb7fd1f7b9757877605cf1e39550627982df50177fe7cd8b66ba352e00000198e076de02
@@ -1,1 +1,0 @@
-{"level":"error","message":"Test error log from diagnostics dashboard","timestamp":"2025-08-25T09:02:23Z"}
