--- .wrangler/state/v3/r2/hassio-proxy-logs/blobs/98a626a18aea09df0c9227dd2fdf72996df05ed64ba9a58302c29cbf1ea8c08500000198e07e5025
@@ -1,1 +1,0 @@
-{"level":"info","message":"Successfully updated diagnostic endpoints to call real Home Assistant APIs","timestamp":"2025-08-25T09:10:31Z"}
