--- .wrangler/state/v3/r2/hassio-proxy-logs/blobs/b16a4d6982bea9656227c6042c971b669201e110c3bdfb0b110776dfbcb40f8500000198e07a7b49
@@ -1,1 +1,0 @@
-{"level":"error","message":"Test error log from diagnostics dashboard","timestamp":"2025-08-25T09:06:20.082Z"}
