+++ public/index.html
@@ -25,12 +25,10 @@
   <style>
     /* Custom animation for the LIVE badge */
     @keyframes pulse-fade {
-
       0%,
       100% {
         opacity: 1;
       }
-
       50% {
         opacity: 0.4;
       }
@@ -39,17 +37,6 @@
     .animate-pulse-fade {
       animation: pulse-fade 2s ease-in-out infinite;
     }
-  </style>
-  }
-
-  50% {
-  opacity: 0.4;
-  }
-  }
-
-  .animate-pulse-fade {
-  animation: pulse-fade 2s cubic-bezier(0.4, 0, 0.6, 1) infinite;
-  }
   </style>
 </head>
